Switch to size_t to fix warnings on Windows builds.

+++ src/sbbs3/wordwrap.c
@@ -37,7 +37,7 @@
 #include "wordwrap.h"
 
 struct prefix {
-	int cols;
+	size_t cols;
 	char *bytes;
 };
 
@@ -194,8 +194,8 @@ static void outbuf_append(char **outbuf, char **outp, char *append, int len, int
  * Length is in bytes and "len" (the number of columns)
  */
 struct section_len {
-	int bytes;
-	int len;
+	size_t bytes;
+	size_t len;
 };
 
 /*
